moved name editor into seperate component
