+++ src/client/components/ApplicationSidebar.jsx
@@ -25,6 +25,7 @@ import { setGridVisible } from "../store/grid";
 import { useSaveSceneAtInterval } from "../hooks/useSaveScene";
 import { useNavigate } from "react-router-dom";
 import { saveFloorplan } from "../store/floorplan";
+import NameEditor from "./ui/NameEditor";
 
 export function ApplicationButtons({ className }) {
   const dispatch = useDispatch();
@@ -162,43 +163,12 @@ function ApplicationSidebar({ className }) {
   const dispatch = useDispatch();
   const { GRID_SIZE, GRID_VISIBLE } = useSelector((state) => state.grid);
   const { isSaving } = useSaveSceneAtInterval(3000);
-  const [edit, setEdit] = useState(false);
-  const [name, setName] = useState("Floorplan Name");
-
-  const handleSave = () => {
-    console.log("saved");
-    setEdit(false);
-  };
-
-  const handleCancel = () => {
-    console.log("cancelled");
-    setEdit(false);
-  };
 
   return (
     <div className={cn("pb-12", className)}>
       <div className="space-y-4 py-4">
         <div className="px-4 py-2">
-          <h2 className="mb-2 px-2 text-lg font-semibold tracking-tight">
-            <div>
-              {edit ? <input placeholder={name}></input> : <h3>{name}</h3>}
-              {!edit && (
-                <Button onClick={() => setEdit(!edit)}>
-                  <Edit className="h-4 w-4" />
-                </Button>
-              )}
-              {edit && (
-                <div>
-                  <Button onClick={handleSave}>
-                    <Save className="h-4 w-4" />
-                  </Button>
-                  <Button onClick={handleCancel}>
-                    <XSquare className="h-4 w-4" />
-                  </Button>
-                </div>
-              )}
-            </div>
-          </h2>
+          <NameEditor />
           <h2 className="mb-2 px-2 text-lg font-semibold tracking-tight">
             Create
           </h2>
